refactor: Move slo-v2 e2e tests to `/test` directory
