+++ test/slov2/slo_request_handling_test.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package v2
+package slov2
 
 import (
 	"strings"
@@ -25,24 +25,24 @@ import (
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/cmd/monaco/integrationtest/v2"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/cmd/monaco/runner"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/featureflags"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/internal/testutils"
 )
 
 func TestIntegrationSloV1AndSloV2(t *testing.T) {
-	configFolder := "test-resources/slo-v1-and-slo-v2/"
+	configFolder := "testdata/slo-v1-and-slo-v2/"
 	manifest := configFolder + "manifest.yaml"
 
 	t.Setenv(featureflags.ServiceLevelObjective.EnvName(), "true")
 
 	t.Run("slo-v1 to slo-v2", func(t *testing.T) {
-		Run(t, configFolder,
-			Options{
-				WithManifestPath(manifest),
-				WithSuffix("slo-v1-to-slo-v2"),
+		v2.Run(t, configFolder,
+			v2.Options{
+				v2.WithSuffix("slo-v1-to-slo-v2"),
 			},
-			func(fs afero.Fs, _ TestContext) {
+			func(fs afero.Fs, _ v2.TestContext) {
 				logOutput := strings.Builder{}
 				cmd := runner.BuildCmdWithLogSpy(testutils.CreateTestFileSystem(), &logOutput)
 				cmd.SetArgs([]string{"deploy", "--verbose", manifest, "--continue-on-error", "--project", "slo-v1-to-slo-v2"})
@@ -57,12 +57,11 @@ func TestIntegrationSloV1AndSloV2(t *testing.T) {
 	})
 
 	t.Run("slo-v2 to slo-v1", func(t *testing.T) {
-		Run(t, configFolder,
-			Options{
-				WithManifestPath(manifest),
-				WithSuffix("slo-v2-to-slo-v1"),
+		v2.Run(t, configFolder,
+			v2.Options{
+				v2.WithSuffix("slo-v2-to-slo-v1"),
 			},
-			func(fs afero.Fs, _ TestContext) {
+			func(fs afero.Fs, _ v2.TestContext) {
 				logOutput := strings.Builder{}
 				cmd := runner.BuildCmdWithLogSpy(testutils.CreateTestFileSystem(), &logOutput)
 				cmd.SetArgs([]string{"deploy", "--verbose", manifest, "--continue-on-error", "--project", "slo-v2-to-slo-v1"})
