+++ test/slov2/testdata/slo-v1-and-slo-v2/manifest.yaml
