+++ test/slov2/testdata/slo-v1-and-slo-v2/slo-v1-to-slo-v2/config.yaml
