+++ test/slov2/testdata/slo-v1-and-slo-v2/slo-v2-to-slo-v1/custom-sli.json
