+++ test/slov2/testdata/v2/manifest.yaml
