+++ test/slov2/testdata/v2/project/config.yaml
