+++ test/slov2/testdata/v2/project/sli-reference.json
